Ignore stderr when running the command
When Rubocop has trouble parsing a file -- or has a bug -- it dumps
information into stderr in the form:

    An error occurred while Lint/DuplicatedKey cop was inspecting $file

This output gets included in the output of the `Command` that gets
parsed in the `Result` and causes a `JSON::ParserError` at
`lib/results.rb:18`.

By ignoring `stderr`, we can avoid this issue and continue on unabated.
Open3 seemed like the cleanest way to do this so I went in that
direction.